requirements.txt
python-telegram-bot==20.7
mercadopago==2.0.14
python-dotenv==1.0.0
opencv-python==4.8.1.78
pillow==10.1.0
numpy==1.26.0